chore: refact to astro build
