+++ .eslintrc.js
@@ -9,15 +9,29 @@ module.exports = {
     "eslint:recommended",
     "@vue/typescript",
     "@vue/prettier",
+    "plugin:astro/recommended",
   ],
-  parser: "vue-eslint-parser",
-  parserOptions: {
-    ecmaVersion: 2020,
-    parser: "@typescript-eslint/parser",
-    sourceType: "module",
-  },
-  rules: {
-    "no-console": process.env.NODE_ENV === "production" ? "warn" : "off",
-    "no-debugger": process.env.NODE_ENV === "production" ? "warn" : "off",
-  },
+  overrides: [
+    {
+    files: ["*.ts", "*.js", "*.vue"],
+    parser: "vue-eslint-parser",
+    parserOptions: {
+      ecmaVersion: 2020,
+      parser: "@typescript-eslint/parser",
+      sourceType: "module",
+    },
+    rules: {
+      "no-console": process.env.NODE_ENV === "production" ? "warn" : "off",
+      "no-debugger": process.env.NODE_ENV === "production" ? "warn" : "off",
+    },
+    },
+    {
+      files: ["*.astro"],
+      parser: "astro-eslint-parser",
+      parserOptions: {
+        parser: "@typescript-eslint/parser",
+        extraFileExtensions: [".astro"],
+      },
+    },
+  ]
 };
