+++ package.json
@@ -6,17 +6,19 @@
   "author": "Daniel Naschberger <daniel@naschi.at>",
   "license": "GPL-3.0",
   "scripts": {
-    "build": "vue-tsc --noEmit && vite build",
-    "lint": "eslint 'src/**/*.{js,ts,vue}' --fix",
+    "dev": "astro dev",
+    "lint": "eslint 'src/**/*.{astro,js,ts,vue}' --fix",
     "lint-no-fix": "eslint 'src/**/*.{js,ts,vue}'",
-    "dev": "vite",
-    "serve": "vite preview"
+    "build": "astro build",
+    "preview": "astro preview"
   },
   "dependencies": {
-    "vue": "^3.2.37",
-    "vue-router": "^4.1.3"
+    "astro": "^1.0.0-rc.3",
+    "vue": "^3.2.37"
   },
   "devDependencies": {
+    "@astrojs/vue": "^0.5.0",
+    "@types/node": "^18.6.3",
     "@typescript-eslint/eslint-plugin": "^5.31.0",
     "@typescript-eslint/parser": "^5.31.0",
     "@vitejs/plugin-vue": "^3.0.1",
@@ -24,13 +26,13 @@
     "@vue/cli-plugin-typescript": "^5.0.8",
     "@vue/eslint-config-prettier": "^7.0.0",
     "@vue/eslint-config-typescript": "^11.0.0",
-    "eslint": "^8.20.0",
+    "eslint": "^8.21.0",
+    "eslint-plugin-astro": "^0.16.0",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-vue": "^9.3.0",
     "prettier": "^2.7.1",
     "prettier-eslint": "^15.0.1",
-    "typescript": "^4.7.4",
-    "vite": "^3.0.4",
-    "vue-tsc": "^0.39.2"
+    "prettier-plugin-astro": "^0.2.0",
+    "typescript": "^4.7.4"
   }
 }
