--- public/404.html
@@ -1,12 +1,0 @@
-<html>
-<head>
-  <title>404</title>
-</head>
-<body style="background-color: #001414; font-family: Avenir, Helvetica, Arial, sans-serif; text-align: center; width: 100%;">
-  <h1 style="color: #d3d3d3; margin: 100px 0 0;">404 - Page not found</h1>
-  <div style="color: #d3d3d3; margin: auto;">
-    <p>The page you are looking for might have been removed or is temporarily unavailable.</p>
-    <p><a href="https://naschi.at" target="_self" style="color: rgb(162, 168, 168);" rel="noopener">https://naschi.at</a></p>
-  </div>
-</body>
-</html>
