--- src/components/GithubIcon.vue
@@ -1,37 +1,0 @@
-// https://github.com/logos
-<template>
-  <a
-    href="https://github.com/naschidaniel"
-    target="_blank"
-    rel="noopener"
-    title="Visit my Github Account"
-    class="logo"
-  >
-    <svg
-      viewBox="0 0 13.3 13.3"
-      version="1.1"
-      xmlns="http://www.w3.org/2000/svg"
-      xmlns:svg="http://www.w3.org/2000/svg"
-    >
-      <g>
-        <path
-          d="m 6.6651575,0.87477154 c -3.2846251,0 -5.94840036,2.67180076 -5.94840036,5.96786046 0,2.636264 1.70439956,4.873203 4.06834536,5.662694 0.2976349,0.05462 0.4061047,-0.129685 0.4061047,-0.287948 0,-0.141783 -0.00518,-0.516922 -0.00828,-1.014799 C 3.5281955,11.563076 3.1790587,10.402475 3.1790587,10.402475 2.9084534,9.7129996 2.5184152,9.5294594 2.5184152,9.5294594 1.9782811,9.1594401 2.5593062,9.1667718 2.5593062,9.1667718 3.1564082,9.2089339 3.4704823,9.7818713 3.4704823,9.7818713 4.0011234,10.693716 4.8629922,10.430317 5.201901,10.277546 5.2559393,9.8921467 5.4096998,9.6291012 5.5795166,9.4799968 4.2585919,9.3290648 2.869726,8.8172728 2.869726,6.5305077 c 0,-0.6517474 0.231909,-1.1840493 0.6124439,-1.6013296 -0.061388,-0.150932 -0.2655018,-0.7576096 0.058386,-1.5793346 0,0 0.4992327,-0.1604653 1.6357441,0.6114441 0.4743877,-0.1322603 0.983486,-0.1982035 1.489282,-0.2007685 0.505444,0.00208 1.0141696,0.068539 1.4892921,0.2007685 1.1357765,-0.7719094 1.6342744,-0.6114441 1.6342744,-0.6114441 0.3246645,0.821725 0.1205198,1.4284026 0.059525,1.5793346 0.3812695,0.4172803 0.6113465,0.9495822 0.6113465,1.6013296 0,2.2926221 -1.3910608,2.7970928 -2.7160021,2.9447225 0.2132751,0.1842775 0.4035477,0.5484288 0.4035477,1.1052918 0,0.797539 -0.00725,1.441217 -0.00725,1.636856 0,0.159728 0.1073726,0.345469 0.409024,0.287211 2.3621244,-0.790945 4.0650534,-3.0260564 4.0650534,-5.661957 0,-3.2960597 -2.663775,-5.96786046 -5.9494873,-5.96786046"
-        ></path>
-      </g>
-    </svg>
-  </a>
-</template>
-
-<style scoped>
-a {
-  align-items: center;
-  display: flex;
-  fill-opacity: 0.8;
-  transition: fill-opacity 0.5s ease-in-out;
-}
-
-a:hover {
-  fill-opacity: 1;
-  transition: fill-opacity 0.5s ease-in-out;
-}
-</style>
