+++ src/components/NaschiLogoIcon.astro
@@ -1,0 +1,65 @@
+<div>
+  <svg
+    viewBox="0 0 66.15 66.15"
+    version="1.1"
+    class="logo"
+    xmlns="http://www.w3.org/2000/svg"
+    xmlns:svg="http://www.w3.org/2000/svg"
+  >
+    <g
+      id="g12"
+      transform="matrix(0.68647093,0,0,-0.68647093,-84.459418,510.27175)"
+    >
+      <g id="g14" clip-path="url(#clipPath16)">
+        <g id="g20" transform="translate(173.6665,724.6719)">
+          <path
+            id="path22"
+            d="m 0,0 c 6.992,0 13.075,-3.952 16.142,-9.737 0.392,-0.739 0.736,-1.506 1.025,-2.3 0.686,-1.887 1.067,-3.92 1.088,-6.036 0.001,-0.063 -0.046,-1.668 -0.133,-2.391 -0.977,-8.087 -7.257,-14.561 -15.251,-15.829 H 2.867 c 0,-0.01 10e-4,-0.018 10e-4,-0.028 0,-0.027 0.005,-0.055 0.005,-0.082 0.011,-2.05 0.419,-4.004 1.139,-5.802 0.007,-0.019 0.013,-0.038 0.021,-0.057 0.013,0.003 0.026,0.005 0.04,0.008 1.036,0.175 2.051,0.412 3.039,0.715 8.998,2.754 15.789,10.592 17.022,20.15 0.104,0.808 0.207,2.904 0.207,3.129 0,1.994 -0.246,3.932 -0.701,5.788 -0.212,0.868 -0.466,1.719 -0.769,2.547 C 19.46,-0.596 10.496,6.081 0,6.081 c -10.384,0 -19.268,-6.54 -22.757,-15.716 2.133,-0.086 4.201,-0.426 6.179,-0.985 C -13.681,-4.358 -7.341,0 0,0"
+            class="upper-part"></path>
+        </g>
+        <g id="g24" transform="translate(149.7969,670.0908)">
+          <path
+            id="path26"
+            d="m 0,0 c -10.055,0 -18.236,8.192 -18.236,18.26 0,9.094 6.682,16.654 15.393,18.034 0.776,0.123 1.571,0.193 2.376,0.214 0.164,0.004 0.326,0.013 0.49,0.013 1.955,0 3.837,-0.313 5.604,-0.883 0.843,-0.273 1.66,-0.605 2.445,-0.992 3.94,-1.943 7.084,-5.264 8.793,-9.332 0.499,-1.187 0.872,-2.438 1.112,-3.735 0.157,-0.859 0.256,-1.739 0.29,-2.635 0.008,-0.227 0.017,-0.454 0.017,-0.684 0,-0.015 0.002,-0.03 0.002,-0.046 0.004,-1.977 0.249,-3.9 0.7,-5.742 0.213,-0.868 0.467,-1.718 0.77,-2.547 0.307,-0.84 0.664,-1.656 1.057,-2.45 C 21.73,5.628 22.875,3.917 24.209,2.372 24.938,1.53 25.721,0.737 26.557,0 Z m 50.088,41.428 c 0.321,-1.655 0.495,-3.361 0.495,-5.107 0,-0.538 -0.02,-1.071 -0.052,-1.601 6.122,-2.952 10.356,-9.221 10.356,-16.46 0,-5.202 -2.191,-9.901 -5.694,-13.23 C 54.448,4.323 53.646,3.674 52.79,3.098 49.883,1.143 46.386,0 42.626,0 41.492,0 40.383,0.109 39.304,0.309 37.31,0.677 35.428,1.37 33.715,2.332 c -3.064,1.72 -5.582,4.297 -7.23,7.405 -0.392,0.739 -0.736,1.506 -1.025,2.3 -0.687,1.887 -1.067,3.92 -1.088,6.037 h 0.004 c -0.001,0.047 -0.008,0.094 -0.009,0.141 0,0.015 -0.002,0.03 -0.002,0.045 0,0.74 -0.038,1.469 -0.104,2.193 -0.309,3.445 -1.337,6.686 -2.936,9.57 -0.433,0.781 -0.909,1.534 -1.423,2.259 -2.771,3.917 -6.666,6.982 -11.214,8.72 -0.736,0.282 -1.489,0.528 -2.257,0.738 -1.942,0.531 -3.978,0.825 -6.079,0.854 -0.109,10e-4 -0.218,0.008 -0.329,0.008 -0.72,0 -1.432,-0.038 -2.137,-0.1 C -14.537,41.416 -24.318,30.961 -24.318,18.26 -24.318,4.839 -13.409,-6.081 0,-6.081 h 42.626 c 7.02,0 13.351,2.991 17.797,7.759 0.628,0.674 1.221,1.381 1.77,2.122 2.998,4.046 4.775,9.049 4.775,14.46 0,10.82 -7.098,20.009 -16.88,23.168"
+            class="lower-part"></path>
+        </g>
+      </g>
+    </g>
+  </svg>
+</div>
+
+<style>
+  div {
+    display: flex;
+    align-items: center;
+    margin: auto;
+    width: 15.2em;
+    height: 15.2em;
+  }
+  .logo {
+    opacity: 1;
+    fill-rule: nonzero;
+    height: 14.5em;
+    margin: auto;
+    opacity: 0.7;
+    stroke: none;
+    transition: all 0.5s ease-in-out;
+    width: 14.5em;
+  }
+  .logo:hover {
+    height: 15.2em;
+    margin: auto;
+    opacity: 1;
+    transition: all 0.5s ease-in-out;
+    width: 15.2em;
+  }
+
+  .upper-part {
+    fill: #2865a2;
+    transition: fill 0.5s ease-in-out;
+  }
+  .lower-part {
+    fill: #008fcf;
+    transition: fill 0.5s ease-in-out;
+  }
+</style>
