--- src/components/NotFound.vue
@@ -1,3 +1,0 @@
-<script setup lang="ts">
-window.location.href = "/404.html";
-</script>
