+++ src/components/OutlineMailIcon.astro
@@ -1,0 +1,33 @@
+---
+// https://heroicons.com/ - by @steveschoger - MIT Licensed - solid mail icon
+---
+
+<a class="box" href="mailto:daniel@naschi.at" title="Write me a message">
+  <svg xmlns="http://www.w3.org/2000/svg" class="icon-mail" viewBox="0 0 20 20">
+    <path
+      d="M2.003 5.884L10 9.882l7.997-3.998A2 2 0 0016 4H4a2 2 0 00-1.997 1.884z"
+    ></path>
+    <path d="M18 8.118l-8 4-8-4V14a2 2 0 002 2h12a2 2 0 002-2V8.118z"></path>
+  </svg>
+</a>
+
+<style>
+  .icon-mail {
+    fill: rgb(201, 201, 201);
+    height: 1.6em;
+    margin-left: 3em;
+    width: 1.6em;
+  }
+
+  a {
+    align-items: center;
+    display: flex;
+    fill-opacity: 0.8;
+    transition: fill-opacity 0.5s ease-in-out;
+  }
+
+  a:hover {
+    fill-opacity: 1;
+    transition: fill-opacity 0.5s ease-in-out;
+  }
+</style>
