+++ src/pages/404.astro
@@ -1,0 +1,42 @@
+<html>
+  <head>
+    <title>naschi.at - 404 - Page not found</title>
+  </head>
+
+  <body>
+    <h1>404 - Page not found</h1>
+    <div class="container">
+      <p>
+        The page you are looking for might have been removed or is temporarily
+        unavailable.
+      </p>
+      <p>
+        <a
+          href="https://naschi.at"
+          target="_self"
+          style="color: rgb(162, 168, 168);"
+          rel="noopener"
+          >https://naschi.at
+        </a>
+      </p>
+    </div>
+    <style>
+      body {
+        background-color: #001414;
+        font-family: Avenir, Helvetica, Arial, sans-serif;
+        text-align: center;
+        width: 100%;
+      }
+
+      h1 {
+        color: #d3d3d3;
+        margin: 100px 0 0;
+      }
+
+      .container {
+        color: #d3d3d3;
+        margin: auto;
+      }
+    </style>
+  </body>
+</html>
